requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
async-timeout==5.0.1
attrs==24.3.0
blessed==1.21.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1648883617327/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.2.1
cycler==0.12.1
datasets==4.4.1
decorator==5.1.1
dill==0.4.0
docker-pycreds==0.4.0
einops==0.8.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filelock==3.16.1
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gpustat==1.1.1
grad-cam==1.5.3
h11==0.14.0
hf-xet==1.1.4
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.33.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1713279365350/work
imagecorruptions==1.1.2
imageio==2.35.1
imgaug==0.4.0
importlib_resources==6.4.0
jenkspy==0.4.1
Jinja2==3.1.5
joblib==1.4.2
kagglehub==0.3.13
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
lightning-utilities==0.11.9
lmdb==1.6.2
lpips==0.1.4
MarkupSafe==3.0.2
matplotlib==3.9.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.18
networkx==3.2.1
nltk==3.9.1
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1701735466804/work
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paconvert==2.0.0
paddlepaddle==2.5.2
pandas==2.2.2
pillow==10.4.0
platformdirs==4.3.8
propcache==0.2.1
protobuf==5.28.3
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.1.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytorch-lightning==2.5.0.post0
pytorch-msssim==1.0.0
pytorch-wavelets==1.3.0
pytz==2024.1
pywavelets @ file:///croot/pywavelets_1705049820073/work
PyYAML==6.0.1
regex==2024.11.6
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1717057054362/work
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
sentry-sdk==2.29.1
setproctitle==1.3.6
shapely==2.0.6
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.2
sniffio==1.3.1
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tokenizers==0.21.0
torch==2.5.1
torchaudio==0.12.0
torchmetrics==1.6.1
torchvision==0.20.1
tqdm==4.66.4
transformers==4.52.4
triton==3.1.0
trojanzoo==2.0.3
ttach==0.0.3
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1718653059220/work
wandb==0.19.11
wcwidth==0.2.13
xxhash==3.6.0
yarl==1.18.3
zipp==3.19.2
